DAT-799 refactor dashboard module
