--- app/src/app/dashboard/components/dashcard/dashcard.component.html
@@ -1,9 +1,0 @@
-<a routerLink="{{linkRoute}}">
-    <div class="dash-card primary">
-        <i class="{{icon}}"></i>
-        <div class="count-section">
-            <div id="count-numbers">{{numbers}}</div>
-            <div id="count-name">{{name}}</div>
-        </div>
-    </div>
-</a>
