--- app/src/app/debug/debug.component.html
@@ -1,126 +1,0 @@
-<h1>Debug mode</h1>
-<hr/>
-<h2>Wrong Routerlink</h2><hr />
-<a [routerLink]="['/foo', 'bar']">Redirect</a>
-
-<h2>HTTP Response</h2>
-<hr/>
-<div class="row">
-
-  <div class="col-md-6">
-    <form [formGroup]="customResponseForm" (ngSubmit)="onCustomResponse()">
-      <div class="form-group row">
-        <label class="col-sm-3 col-form-label">Trigger HTTP Response</label>
-        <div class="col-sm-9">
-          <input type="number" formControlName="customStatus" min="400" max="599" value="400" class="form-control">
-        </div>
-      </div>
-      <div class="form-group row">
-        <label class="col-sm-3 col-form-label">Custom Message</label>
-        <div class="col-sm-9">
-          <input type="text" formControlName="customMessage" class="form-control">
-        </div>
-      </div>
-      <div class="form-group row">
-        <div class="col-sm-10">
-          <button type="submit" class="btn btn-primary" [disabled]="customResponseForm.invalid">Send</button>
-        </div>
-      </div>
-    </form>
-  </div>
-  <div class="col-md-6">
-    Response: {{customResponse | json}}
-  </div>
-</div>
-
-<table class="table table-striped">
-  <caption>4xx Client errors</caption>
-  <thead>
-  <tr>
-    <th scope="col">Action</th>
-    <th scope="col">Status code</th>
-    <th scope="col">Message</th>
-    <th scope="col">Expected behavior</th>
-    <th scope="col" colspan="2">Response</th>
-  </tr>
-  </thead>
-  <tbody>
-  <tr>
-    <th scope="row">
-      <button (click)="onStartHttpResponse(400)" type="button" class="btn btn-sm btn-block btn-danger">
-        <i class="fas fa-caret-right"></i>
-      </button>
-    </th>
-    <td>400</td>
-    <td>Bad Request</td>
-    <td>Error modal</td>
-    <td id="http_400_response" colspan="2"></td>
-  </tr>
-  <tr>
-    <th scope="row">
-      <button (click)="onStartHttpResponse(401)" type="button" class="btn btn-sm btn-block btn-danger">
-        <i class="fas fa-caret-right"></i>
-      </button>
-    </th>
-    <td>401</td>
-    <td>Unauthorized</td>
-    <td>Logout/Redirect</td>
-    <td id="http_401_response" colspan="2"></td>
-  </tr>
-  <tr>
-    <th scope="row">
-      <button (click)="onStartHttpResponse(403)" type="button" class="btn btn-sm btn-block btn-danger">
-        <i class="fas fa-caret-right"></i>
-      </button>
-    </th>
-    <td>403</td>
-    <td>Forbidden</td>
-    <td>Redirect error page</td>
-    <td id="http_403_response" colspan="2"></td>
-  </tr>
-  <tr>
-    <th scope="row">
-      <button (click)="onStartHttpResponse(404)" type="button" class="btn btn-sm btn-block btn-danger">
-        <i class="fas fa-caret-right"></i>
-      </button>
-    </th>
-    <td>404</td>
-    <td>Not Found</td>
-    <td>Error modal</td>
-    <td id="http_404_response" colspan="2"></td>
-  </tr>
-  <tr>
-    <th scope="row">
-      <button (click)="onStartHttpResponse(405)" type="button" class="btn btn-sm btn-block btn-danger">
-        <i class="fas fa-caret-right"></i>
-      </button>
-    </th>
-    <td>405</td>
-    <td>Method Not Allowed</td>
-    <td>Error modal</td>
-    <td id="http_405_response" colspan="2"></td>
-  </tr>
-  <tr>
-    <th scope="row">
-      <button (click)="onStartHttpResponse(406)" type="button" class="btn btn-sm btn-block btn-danger">
-        <i class="fas fa-caret-right"></i>
-      </button>
-    </th>
-    <td>406</td>
-    <td>Not Acceptable</td>
-    <td>Error modal</td>
-    <td id="http_406_response" colspan="2"></td>
-  </tr>
-  <tr>
-    <th scope="row">
-      <button (click)="onStartHttpResponse(410)" type="button" class="btn btn-sm btn-block btn-danger">
-        <i class="fas fa-caret-right"></i>
-      </button>
-    </th>
-    <td>410</td>
-    <td>Page Gone</td>
-    <td>Error modal</td>
-    <td id="http_410_response" colspan="2"></td>
-  </tr>
-  </tbody>
-</table>
