--- app/src/app/debug/debug.service.ts
@@ -1,51 +1,0 @@
-/*
-* DATAGERRY - OpenSource Enterprise CMDB
-* Copyright (C) 2024 becon GmbH
-*
-* This program is free software: you can redistribute it and/or modify
-* it under the terms of the GNU Affero General Public License as
-* published by the Free Software Foundation, either version 3 of the
-* License, or (at your option) any later version.
-*
-* This program is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-* GNU Affero General Public License for more details.
-*
-* You should have received a copy of the GNU Affero General Public License
-* along with this program.  If not, see <https://www.gnu.org/licenses/>.
-*/
-import { Injectable } from '@angular/core';
-
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
-
-import { ApiCallService, ApiServicePrefix } from '../services/api-call.service';
-/* ------------------------------------------------------------------------------------------------------------------ */
-
-
-@Injectable({
-  providedIn: 'root'
-})
-export class DebugService<T = any> implements ApiServicePrefix {
-  public servicePrefix: string = 'debug';
-
-  public constructor(private api: ApiCallService) {
-  }
-
-  public getResponse(statusCode: number): Observable<T> {
-    return this.api.callGet<T>(`${ this.servicePrefix }/error/${ statusCode }`).pipe(
-      map((apiResponse) => {
-        return apiResponse.body;
-      })
-    );
-  }
-
-  public getResponseWithMessage(statusCode: number, customMessage: string): Observable<T> {
-    return this.api.callGet<T>(`${ this.servicePrefix }/error/${ statusCode }/${customMessage}`).pipe(
-      map((apiResponse) => {
-        return apiResponse.body;
-      })
-    );
-  }
-}
