+++ app/src/app/main/main.module.ts
@@ -19,7 +19,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MainRoutingModule } from './main-routing.module';
-import { DashboardModule } from '../dashboard/dashboard.module';
+import { DashboardModule } from '../modules/dashboard/dashboard.module';
 import { NgxIndexedDBModule } from 'ngx-indexed-db';
 import { userSettingsDBConfig } from '../management/user-settings/user-settings.module';
 import { UserSettingsDBService } from '../management/user-settings/services/user-settings-db.service';
