+++ app/src/app/modules/dashboard/dashboard.component.scss
